backend/requirements/base.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0

# Authentication
djangorestframework-simplejwt==5.3.0

# CORS
django-cors-headers==4.3.0

# Filtering
django-filter==23.3

# Image handling
Pillow==10.1.0

# Utilities
python-decouple==3.8